add scrollable containers

+++ string.cpp
@@ -2,8 +2,8 @@
 
 std::vector<std::string> split(std::string& text, const std::string& delimiter) {
     std::vector<std::string> tokens;
-    int startPos = 0;
-    int endPos = 0;
+    size_t startPos = 0;
+    size_t endPos = 0;
     std::string token;
     while ((endPos = text.find(delimiter, startPos)) != std::string::npos) {
         tokens.push_back(text.substr(startPos, endPos - startPos));
